src: consolidate project source
