+++ Makefile
@@ -1,6 +1,6 @@
 include Makefile.inc
 
-SUBDIRS = libutil libini libcommon motiond etc # dts
+SUBDIRS = libini src etc # dts
 
 all:
