+++ src/Makefile
@@ -1,17 +1,19 @@
 include ../Makefile.inc
 
 PROG = gem-motiond
+
 CFLAGS = -Wall -D_GNU_SOURCE=1 -I$(abs_topdir) \
-	$(ZMQ_CFLAGS)
+	 -DCONFIG_FILENAME=\"$(prefix)/etc/gem.config\"
 
 LIBS  = -L$(abs_topdir)/libutil -lutil \
-	-L$(abs_topdir)/libcommon -lcommon \
 	-L$(abs_topdir)/libini -lini \
-	$(ZMQ_LIBS) -lpthread -lev -lm -lrt 
+	-lpthread -lev -lm -lrt
+
+OBJS = configfile.o xzmalloc.o log.o gpio.o hpad.o guide.o motion.o daemon.o
 
 all: $(PROG)
 
-$(PROG): hpad.o guide.o motion.o daemon.o
+$(PROG): $(OBJS)
 	$(CC) -o $@ $^ $(LIBS)
 
 install: $(PROG)
