+++ src/guide.h
